+++ TeamProject/Assets/Sprite.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79067cf07ecd8a145917f825f6bdd103
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
