+++ TeamProject/Assets/script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db06467c2ef84ff47bee2f0845c5e5b9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
