+++ TeamProject/Assets/script/NPC_Script.cs.meta
