+++ TeamProject/Assets/script/eettt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f5fccc30bf8e8c4a8a178121c272244
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
