+++ TeamProject/Assets/script/start_buttn.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+public class start_buttn : MonoBehaviour
+{
+    public void playGame(){
+
+SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
+}
+
+public void exit(){
+Application.Quit();
+
+
+}
+
+
+
+}
+
