+++ TeamProject/Assets/script/start_buttn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 209697950de113c4db440af10e30d44c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
