+++ TeamProject/Packages/manifest.json
@@ -1,15 +1,15 @@
 {
   "dependencies": {
-    "com.unity.collab-proxy": "1.17.7",
-    "com.unity.feature.2d": "1.0.0",
-    "com.unity.ide.rider": "3.0.18",
-    "com.unity.ide.visualstudio": "2.0.17",
+    "com.unity.collab-proxy": "2.1.0",
+    "com.unity.feature.2d": "2.0.0",
+    "com.unity.ide.rider": "3.0.25",
+    "com.unity.ide.visualstudio": "2.0.21",
     "com.unity.ide.vscode": "1.2.5",
-    "com.unity.test-framework": "1.1.31",
+    "com.unity.test-framework": "1.1.33",
     "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.6.4",
+    "com.unity.timeline": "1.6.5",
     "com.unity.ugui": "1.0.0",
-    "com.unity.visualscripting": "1.8.0",
+    "com.unity.visualscripting": "1.9.1",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
