+++ TeamProject/ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,12 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/Main_Game_Menu.unity
+    guid: f67a1887dfcb5c64092d1d9e6d134090
+  - enabled: 1
+    path: Assets/Scenes/change_skin.unity
+    guid: 69459080a0a4b3841a33c8210854f7b0
+  - enabled: 1
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   m_configObjects: {}
