+++ TeamProject/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.17f1
-m_EditorVersionWithRevision: 2021.3.17f1 (3e8111cac19d)
+m_EditorVersion: 2021.3.31f1
+m_EditorVersionWithRevision: 2021.3.31f1 (3409e2af086f)
